requirements.txt
# Core libraries
crewai==0.102.0
python-dotenv==1.0.0
setuptools>=75.0.0

# Telegram integration
python-telegram-bot>=20.4

# API tools
google-search-results>=2.4.2
serpapi>=0.1.5

# LangChain (for custom tools in CrewAI)
langchain>=0.3.19
langchain-community>=0.3.18

# LLM Providers
anthropic>=0.19.0  # For Claude AI models

# Time Agent dependencies
ephem>=4.1.4
pytz>=2023.3

# Calendar Agent dependencies
google-api-python-client>=2.107.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.1

# You may add more dependencies here as you create additional agents 